+++ order.go
@@ -10,13 +10,13 @@ const (
 )
 
 type Order struct {
-	Side      Side   `json:"side"`
-	Price     uint64 `json:"price"`
-	Size      uint64 `json:"size"`
-	Timestamp int64  `json:"timestamp"`
+	Side      Side    `json:"side"`
+	Price     float64 `json:"price"`
+	Size      float64 `json:"size"`
+	Timestamp int64   `json:"timestamp"`
 }
 
-func newOrder(side Side, size uint64) *Order {
+func newOrder(side Side, size float64) *Order {
 	return &Order{
 		Side:      side,
 		Size:      size,
